+++ qml/pages/MainPage.qml
@@ -2,25 +2,29 @@ import QtQuick 2.0
 import Sailfish.Silica 1.0
 import QtMultimedia 5.6
 import Nemo.Notifications 1.0
-
+import QtFeedback 5.0
+import "../components"
 import "../js/slack.js" as Slack
 
 Page {
     id: mainPage
     property string notificationChannel
     property int unreadCount
-    property string leSlackToken: ""
-    property variant privateChannels: ['','', '']
-    property variant publicChannels: ['', '']
+    property string inlineSlackToken: ""
     property string defaultNotificationSound: Qt.resolvedUrl("../sounds/slack-notification.wav")
 
     allowedOrientations: Orientation.All
-
 
     SoundEffect {
         id: alertSound
         volume: 0.6
         source: "slack-notification.wav"
+    }
+    HapticsEffect {
+        id: alertVibration
+        duration: 500
+        intensity: 0.2
+        running: appSettings.vibration
     }
 
     Notification {
@@ -42,11 +46,14 @@ Page {
         running: appSettings.onDuty
         repeat: true
         onTriggered: {
-            for (var i = 0; i < mainPage.publicChannels.length; i++) {
-                if(Slack.get_unread_private(mainPage.leSlackToken, mainPage.privateChannels[i]) > 0) {
+            for (var i = 0; i < channelsModel.count; ++i) {
+                var channel = channelsModel.get(i)
+                var unreadMessages = Slack.get_unread(mainPage.inlineSlackToken, channel.channel_id, channel.type)
+                if( unreadMessages > 0) {
                     alertSound.play()
-                    mainPage.notificationChannel = Slack.get_private_channel_info(mainPage.leSlackToken, mainPage.privateChannels[i])
-                    mainPage.unreadCount = Slack.get_unread_private(mainPage.leSlackToken,mainPage.privateChannels[i])
+                    alertVibration.start()
+                    mainPage.notificationChannel = channel.name
+                    mainPage.unreadCount = unreadMessages
                     notification.republish()
                 }
             }
@@ -79,74 +86,73 @@ Page {
                 title: qsTr("Slack Notifish")
             }
 
-            SectionHeader { text: qsTr("Public Channels") }
+            SectionHeader { text: qsTr("Channels") }
             ColumnView {
-                 model: publicListModel
+                 model: channelsModel
                  itemHeight: Theme.itemSizeSmall
 
                  delegate: ListItem {
-                      Label {
-                        text: "#" + model.name
-                        font.pixelSize: Theme.fontSizeMedium
-                        width: parent.width
-                        color: Theme.primaryColor
-                        horizontalAlignment: Text.AlignLeft
-                        x: Theme.paddingMedium
-                      }
-
-                      Label {
-                        text: model.channel_id
-                        font.pixelSize: Theme.fontSizeMedium
-                        width: parent.width
-                        color: Theme.secondaryColor
-                        horizontalAlignment: Text.AlignRight
-                        x: Theme.paddingMedium
-                      }
-                 }
-            }
-
-            SectionHeader { text: qsTr("Private Channels") }
-            ColumnView {
-                 model: privateListModel
-                 itemHeight: Theme.itemSizeSmall
-
-                 delegate: ListItem {
+                     onClicked: {
+                        pageStack.push(Qt.resolvedUrl('ChannelInfo.qml'), {channelID: model.channel_id, channelName: model.name, channelType: model.type})
+                     }
+                     Row {
+                         id: row
+                         anchors.verticalCenter: parent.verticalCenter
+                         x: Theme.paddingSmall * (Screen.sizeCategory >= Screen.Small ? 2 : 1)
+                         spacing: Theme.paddingSmall
+                     IconButton {
+                        icon.source: setChannelIcon()
+                        anchors.verticalCenter: parent.verticalCenter
+                        function setChannelIcon()
+                        {
+                            var imageIcon
+                            if(model.type === "public")
+                            {
+                                imageIcon = "image://theme/icon-m-chat"
+                            }
+                            if(model.type === "private")
+                            {
+                                imageIcon = "image://theme/icon-m-device-lock"
+                            }
+                            return imageIcon
+                        }
+                     }
                           Label {
                             text: "#" + model.name
                             font.pixelSize: Theme.fontSizeMedium
-                            width: parent.width
-                            horizontalAlignment: Text.AlignLeft
                             color: Theme.primaryColor
-                            x: Theme.paddingMedium
                           }
                           Label {
                             text: model.channel_id
-                            font.pixelSize: Theme.fontSizeMedium
-                            width: parent.width
+                            font.pixelSize: Theme.fontSizeTiny
                             color: Theme.secondaryColor
-                            horizontalAlignment: Text.AlignRight
-                            x: Theme.paddingMedium
                           }
+                     }
                  }
+            }
+            Repeater {
+                model: [
+                    qsTr("No Channels, go to settings")
+                ]
+                Label {
+                  id: infoViewLabel
+                  text: modelData
+                  font.pixelSize: Theme.fontSizeSmall
+                  x: Theme.paddingLarge
+                  wrapMode: Text.Wrap
+                  color: Theme.secondaryColor
+                  visible: checkChannels()
+                  function checkChannels() {
+                      if(channelsModel.count == 0) {
+                        return true
+                      } else {
+                        return false
+                      }
+                  }
+                }
             }
 
         }
-    }
 
-    ListModel {
-        id: privateListModel
-    }
-    ListModel {
-        id: publicListModel
-    }
-    Component.onCompleted: {
-        for (var i = 0; i < mainPage.privateChannels.length; i++) {
-            privateListModel.append({"name":Slack.get_private_channel_info(mainPage.leSlackToken, mainPage.privateChannels[i]), "channel_id":mainPage.privateChannels[i]})
-        }
-        for (var d = 0; d < mainPage.publicChannels.length; d++) {
-            publicListModel.append({"name":Slack.get_public_channel_info(mainPage.leSlackToken, mainPage.publicChannels[d]), "channel_id":mainPage.publicChannels[d]})
-        }
     }
 }
-
-