+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.3
+- Add better UI
+- Add Channels in Settings
+- Use Database for store the channels
+- New App Icon
+
 ## v0.2
 - Add own Notification Sound